refactor: extract DCA state to dca.rs
Move DcaState and its implementation out of lib.rs into a dedicated
dca.rs module. Fields and methods are pub(crate) since they are accessed
from the execution helpers.
